+++ src/layers/jacobian_stabilization.jl
@@ -1,65 +1,65 @@
-import CUDA
-import ChainRulesCore as CRC
-import Lux
-import Random
-import Statistics
+# import CUDA
+# import ChainRulesCore as CRC
+# import Lux
+# import Random
+# import Statistics
 
-gaussian_like(rng::Random.AbstractRNG, x::AbstractArray) = randn(rng, eltype(x), size(x))
-function gaussian_like(rng::Random.AbstractRNG, x::CUDA.CuArray)
-  return CUDA.randn(rng, eltype(x), size(x))
-end
+# gaussian_like(rng::Random.AbstractRNG, x::AbstractArray) = randn(rng, eltype(x), size(x))
+# function gaussian_like(rng::Random.AbstractRNG, x::CUDA.CuArray)
+#   return CUDA.randn(rng, eltype(x), size(x))
+# end
 
-"""
-    estimate_jacobian_trace(::Val{mode}, model::Lux.AbstractExplicitLayer, ps,
-                            st::NamedTuple, z::AbstractArray, x::AbstractArray,
-                            rng::Random.AbstractRNG)
+# """
+#     estimate_jacobian_trace(::Val{mode}, model::Lux.AbstractExplicitLayer, ps,
+#                             st::NamedTuple, z::AbstractArray, x::AbstractArray,
+#                             rng::Random.AbstractRNG)
 
-Estimates the trace of the jacobian matrix wrt `z`.
+# Estimates the trace of the jacobian matrix wrt `z`.
 
-## Arguments
+# ## Arguments
 
-  - `mode`: Should be either `:reverse_mode` or `:finite_diff`. `:reverse_mode` uses
-    `Zygote.jl` while `:finite_diff` uses Finite Differencing. If using inside a gradient
-    call, `:reverse_mode` might not work, and `:finite_diff` is the recommended approach.
-  - `model`: A `Lux` Neural Network mapping 2 equal sized arrays to a same sized array. This
-    convention is not checked, and if violated will lead to errors.
-  - `ps`: Parameters of `model`.
-  - `st`: States of `model`.
-  - `z`: Input wrt the Jacobian is computed.
-  - `x`: Other Input to `model`.
-  - `rng`: PRNG. Note that this object is mutated by this function.
+#   - `mode`: Should be either `:reverse_mode` or `:finite_diff`. `:reverse_mode` uses
+#     `Zygote.jl` while `:finite_diff` uses Finite Differencing. If using inside a gradient
+#     call, `:reverse_mode` might not work, and `:finite_diff` is the recommended approach.
+#   - `model`: A `Lux` Neural Network mapping 2 equal sized arrays to a same sized array. This
+#     convention is not checked, and if violated will lead to errors.
+#   - `ps`: Parameters of `model`.
+#   - `st`: States of `model`.
+#   - `z`: Input wrt the Jacobian is computed.
+#   - `x`: Other Input to `model`.
+#   - `rng`: PRNG. Note that this object is mutated by this function.
 
-## Returns
+# ## Returns
 
-Stochastic Estimate of the trace of the Jacobian.
-"""
-function estimate_jacobian_trace(::Val{:reverse_mode}, model::Lux.AbstractExplicitLayer, ps,
-                                 st::NamedTuple, z::AbstractArray, x::AbstractArray,
-                                 rng::Random.AbstractRNG)
-  _, back = Zygote.pullback(u -> model((u, x), ps, st)[1], z)
-  vjp_z = back(gaussian_like(rng, x))[1]
-  return Statistics.mean(abs2, vjp_z)
-end
+# Stochastic Estimate of the trace of the Jacobian.
+# """
+# function estimate_jacobian_trace(::Val{:reverse_mode}, model::Lux.AbstractExplicitLayer, ps,
+#                                  st::NamedTuple, z::AbstractArray, x::AbstractArray,
+#                                  rng::Random.AbstractRNG)
+#   _, back = Zygote.pullback(u -> model((u, x), ps, st)[1], z)
+#   vjp_z = back(gaussian_like(rng, x))[1]
+#   return Statistics.mean(abs2, vjp_z)
+# end
 
-function estimate_jacobian_trace(::Val{:finite_diff}, model::Lux.AbstractExplicitLayer, ps,
-                                 st::NamedTuple, z::AbstractArray, x::AbstractArray,
-                                 rng::Random.AbstractRNG)
-  f = u -> model((u, x), ps, st)[1]
-  res = convert(eltype(z), 0)
-  epsilon = cbrt(eps(typeof(res)))
-  _epsilon = inv(epsilon)
-  f0 = f(z)
-  v = gaussian_like(rng, x)
+# function estimate_jacobian_trace(::Val{:finite_diff}, model::Lux.AbstractExplicitLayer, ps,
+#                                  st::NamedTuple, z::AbstractArray, x::AbstractArray,
+#                                  rng::Random.AbstractRNG)
+#   f = u -> model((u, x), ps, st)[1]
+#   res = convert(eltype(z), 0)
+#   epsilon = cbrt(eps(typeof(res)))
+#   _epsilon = inv(epsilon)
+#   f0 = f(z)
+#   v = gaussian_like(rng, x)
 
-  for idx in eachindex(z)
-    _z = z[idx]
-    CRC.ignore_derivatives() do
-      return z[idx] = z[idx] + epsilon
-    end
-    res = res + abs2(sum((f(z) .- f0) .* _epsilon .* v)) / length(v)
-    CRC.ignore_derivatives() do
-      return z[idx] = _z
-    end
-  end
-  return res
-end
+#   for idx in eachindex(z)
+#     _z = z[idx]
+#     CRC.ignore_derivatives() do
+#       return z[idx] = z[idx] + epsilon
+#     end
+#     res = res + abs2(sum((f(z) .- f0) .* _epsilon .* v)) / length(v)
+#     CRC.ignore_derivatives() do
+#       return z[idx] = _z
+#     end
+#   end
+#   return res
+# end
